% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatasets.R
\docType{data}
\name{Depression_column}
\alias{Depression_column}
\title{Supplementary data to be added to Depression data}
\format{
\describe{
   \item{6week.1}{A value of 1 indicates that depression is "Abnormal" after 6 weeks}
   \item{6week.2}{A value of 1 indicates that depression is "Normal" after 6 weeks}
   \item{sex.1}{A value of 1 indicates that the gender is "Male"}
   \item{sex.2}{A value of 1 indicates that the gender is "Female"}
}
}
\usage{
data(Depression_column)
}
\description{
Artificially created data to add the degree and sex of depression after 6 weeks.
}
\details{
{
As supplementary variables, for 800 patients, response at fourth time point (after 6 weeks) and gender that could affect depression were added to the columns. \cr
Indicator matrix of 800 rows and 4 dummy variables.
\tabular{rccccc}{
\strong{ }  \tab \strong{6weeks.1}  \tab \strong{6weeks.2} \tab \strong{sex.1} \tab \strong{sex.2} \cr
1       \tab 0        \tab 1        \tab 1      \tab 0     \cr
2       \tab 0        \tab 1        \tab 1      \tab 0     \cr
3       \tab 0        \tab 1        \tab 1      \tab 0     \cr
4       \tab 0        \tab 1        \tab 1      \tab 0     \cr
5       \tab 0        \tab 1        \tab 1      \tab 0     \cr
6       \tab 0        \tab 1        \tab 1      \tab 0     \cr
7       \tab 0        \tab 1        \tab 1      \tab 0     \cr
8       \tab 0        \tab 1        \tab 1      \tab 0     \cr
9       \tab 0        \tab 1        \tab 1      \tab 0     \cr
10      \tab 0        \tab 1        \tab 1      \tab 0     \cr
\code{.}\tab \code{.} \tab \code{.} \tab\code{.}\tab\code{.}\cr
\code{.}\tab \code{.} \tab \code{.} \tab\code{.}\tab\code{.}\cr
\code{.}\tab \code{.} \tab \code{.} \tab\code{.}\tab\code{.}\cr
796     \tab 0        \tab 1        \tab 0      \tab 1     \cr
797     \tab 0        \tab 1        \tab 0      \tab 1     \cr
798     \tab 0        \tab 1        \tab 0      \tab 1     \cr
799     \tab 0        \tab 1        \tab 0      \tab 1     \cr
800     \tab 0        \tab 1        \tab 0      \tab 1     \cr
 }
}
}
\keyword{datasets}
