% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatasets.R
\docType{data}
\name{Depression_row}
\alias{Depression_row}
\title{Supplementary data to be added to Depression data}
\format{
\describe{
   \item{Diagnosis.1}{A value of 1 indicates the "Severe" of depression}
   \item{Diagnosis.2}{A value of 1 indicates the "Mild" of depression}
   \item{Drug.1}{A value of 1 indicates that the drug being taken is a “New drug”}
   \item{Drug.2}{A value of 1 indicates that the drug being taken is a “Standard drug”}
   \item{1week.1}{A value of 1 indicates that depression is "Abnormal" after 1 week}
   \item{1week.2}{A value of 1 indicates that depression is "Normal" after 1 week}
   \item{2week.1}{A value of 1 indicates that depression is "Abnormal" after 2 weeks}
   \item{2week.2}{A value of 1 indicates that depression is "Normal" after 2 weeks}
   \item{4week.1}{A value of 1 indicates that depression is "Abnormal" after 4 weeks}
   \item{4week.2}{A value of 1 indicates that depression is "Normal" after 4 weeks}
}
}
\usage{
data(Depression_row)
}
\description{
Artificially generated data to add a placebo effect that affects the degree of depression.
}
\details{
{
Supplementary 100 objects are patients who take placebo. \cr
Indicator matrix of 100 rows and 10 dummy variables
 \tabular{rcccccccccc}{
\strong{ }\tab \strong{Diag.1}\tab \strong{Diag.2}\tab \strong{Drug.1}\tab \strong{Drug.2}\tab \strong{.}\tab \strong{.}\tab \strong{.}\tab \strong{2week.2} \tab \strong{4week.1} \tab \strong{4week.2} \cr
1       \tab 1       \tab 0      \tab 0      \tab 0      \tab \code{.}\tab\code{.}\tab\code{.}\tab 0    \tab 1    \tab 0 \cr
2       \tab 1       \tab 0      \tab 0      \tab 0      \tab        \tab     \tab         \tab 0    \tab 1    \tab 0 \cr
3       \tab 1       \tab 0      \tab 0      \tab 0      \tab\code{.}\tab\code{.} \tab\code{.}         \tab 0    \tab 1    \tab 0 \cr
4       \tab 1       \tab 0      \tab 0      \tab 0      \tab        \tab     \tab         \tab 0    \tab 0    \tab 1 \cr
5       \tab 1       \tab 0      \tab 0      \tab 0      \tab\code{.}\tab\code{.} \tab\code{.}         \tab 0    \tab 0    \tab 1 \cr
6       \tab 1       \tab 0      \tab 0      \tab 0      \tab        \tab     \tab         \tab 0    \tab 0    \tab 1 \cr
7       \tab 1       \tab 0      \tab 0      \tab 0      \tab\code{.}\tab\code{.} \tab\code{.}         \tab 0    \tab 0    \tab 1 \cr
8       \tab 1       \tab 0      \tab 0      \tab 0      \tab        \tab     \tab         \tab 0    \tab 0    \tab 1 \cr
9       \tab 1       \tab 0      \tab 0      \tab 0      \tab\code{.}\tab\code{.} \tab\code{.}         \tab 1    \tab 1    \tab 0 \cr
10      \tab 1       \tab 0      \tab 0      \tab 0      \tab        \tab     \tab         \tab 1    \tab 1    \tab 0 \cr
\code{.}\tab\code{.} \tab\code{.}\tab\code{.}\tab\code{.}\tab\code{.}\tab \code{} \tab\code{}\tab\code{.}\tab\code{.} \tab\code{.}\cr
\code{.}\tab\code{.} \tab\code{.}\tab\code{.}\tab\code{.}\tab\code{}\tab \code{.} \tab\code{}\tab\code{.}\tab\code{.} \tab\code{.}\cr
\code{.}\tab\code{.} \tab\code{.}\tab\code{.}\tab\code{.}\tab\code{}\tab \code{} \tab\code{.}\tab\code{.}\tab\code{.} \tab\code{.}\cr
96      \tab 0       \tab 1      \tab 0      \tab 0      \tab\code{.}\tab\code{.}\tab\code{.}\tab 1    \tab 0    \tab 1 \cr
97      \tab 0       \tab 1      \tab 0      \tab 0      \tab        \tab     \tab         \tab 1    \tab 0    \tab 1 \cr
98      \tab 0       \tab 1      \tab 0      \tab 0      \tab\code{.}\tab\code{.}\tab\code{.}\tab 1    \tab 0    \tab 1 \cr
99      \tab 0       \tab 1      \tab 0      \tab 0      \tab        \tab     \tab         \tab 1    \tab 0    \tab 1 \cr
100     \tab 0       \tab 1      \tab 0      \tab 0      \tab\code{.}\tab\code{.}\tab\code{.}\tab 1    \tab 0    \tab 1 \cr
 }
}
}
\keyword{datasets}
