% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indicator.R
\name{indicator}
\alias{indicator}
\title{Indicator matrix}
\usage{
indicator(x)
}
\arguments{
\item{x}{A data frame of categorical data coded in numbers.}
}
\value{
Dummy_variables
}
\description{
Convert values of categorical variables into indicator matrix
}
\examples{
## Long form
data(PTSD)
PTSD <- as.data.frame(PTSD)
# Transform a string or continuous class variable into factor
PTSD[,2:4] <- apply(PTSD[,2:4], 2, function(x) ifelse(x >= 3, 1, 0))
PTSD[,5] <-  ifelse(PTSD[,5] >= 6 , 1, 0)
PTSD <- data.frame(lapply(PTSD[,-1], function(x) as.factor(x)))
indicator(PTSD)


## Wide form
data(Depression)
str(Depression)
indicator(Depression[,-1])



}
\keyword{Indicator}
