% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vlda.R
\name{print.vlda}
\alias{print.vlda}
\title{Print a vlda object}
\usage{
\method{print}{vlda}(x, ...)
}
\arguments{
\item{x}{A vlda object to print}

\item{...}{Other arguments not used by this method}
}
\value{
Invisibly returns the result of vlda, which is a list of components that contain the data itself, information etc.
}
\description{
Print method for vlda
}
