% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatasets.R
\docType{data}
\name{PTSD_column}
\alias{PTSD_column}
\title{Supplementary data to be added to PTSD data}
\format{
\describe{
   \item{Drinking.0}{A value of 1 indicates that the degree of drinking is low}
   \item{Drinking.1}{A value of 1 indicates that the degree of drinking is high}
}
}
\usage{
data(PTSD_column)
}
\description{
Artificially created data to add drinking level to PTSD data.
}
\details{
{
The degree of drinking (low, high) that can affect PTSD is added to the columns corresponding to the first to third time points for 316 patients. \cr
Indicator matrix of 948 rows and 2 dummy variables.
\tabular{rcc}{
\strong{ } \tab  \strong{Drinking.0}  \tab \strong{Drinking.0}\cr
1         \tab 1        \tab 0    \cr
2         \tab 1        \tab 0    \cr
3         \tab 1        \tab 0    \cr
4         \tab 0        \tab 1    \cr
5         \tab 1        \tab 0    \cr
6         \tab 1        \tab 0    \cr
7         \tab 0        \tab 1    \cr
8         \tab 0        \tab 1    \cr
9         \tab 0        \tab 1    \cr
10        \tab 1        \tab 0    \cr
\code{.}  \tab \code{.} \tab \code{.} \cr
\code{.}  \tab \code{.} \tab \code{.} \cr
\code{.}  \tab \code{.} \tab \code{.} \cr
944       \tab 1        \tab 0    \cr
945       \tab 1        \tab 0    \cr
946       \tab 1        \tab 0    \cr
947       \tab 1        \tab 0    \cr
948       \tab 1        \tab 0    \cr
 }
}
}
\keyword{datasets}
