% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatasets.R
\docType{data}
\name{PTSD_row}
\alias{PTSD_row}
\title{Supplementary data to be added to PTSD data}
\format{
\describe{
   \item{control.0}{A value of 1 indicates low control}
   \item{control.1}{A value of 1 indicates high control}
   \item{problems.0}{A value of 1 indicates that the degree of problems is low}
   \item{problems.1}{A value of 1 indicates that the degree of problems is high}
   \item{stress.0}{A value of 1 indicates that the degree of stress is low}
   \item{stress.1}{A value of 1 indicates that the degree of stress is high}
   \item{cohesion.0}{A value of 1 indicates that the degree of stress is low}
   \item{cohesion.1}{A value of 1 indicates that the degree of stress is high}
   \item{time.1}{Zero vector (All elements is zero)}
   \item{time.2}{Zero vector (All elements is zero)}
   \item{time.3}{Zero vector (All elements is zero)}
   \item{ptsd.0}{A value of 1 indicates a low post-traumatic stress disorder}
   \item{ptsd.1}{A value of 1 indicates a high post-traumatic stress disorder}
}
}
\usage{
data(PTSD_row)
}
\description{
Artificially created data to add variables after 18 months to PTSD data.
}
\details{
{
This data is a long form of control, problem, stress, stress, stress and PTSD added to the row, and is intended for 316 patients after 18 months. \cr
Indicator matrix of 316 rows and 13 dummy variables.
\tabular{rcccccccccccc}{
\strong{ } \tab \strong{control.0}  \tab \strong{control.1}  \tab \strong{.} \tab \strong{.} \tab \strong{.} \tab \strong{time.1} \tab \strong{time.2} \tab \strong{time.3} \tab \strong{ptsd.0} \tab \strong{ptsd.1}\cr
1       \tab 0       \tab 1        \tab \code{.} \tab \code{.} \tab \code{.} \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
2       \tab 0       \tab 1        \tab \code{ } \tab \code{} \tab \code{ } \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
3       \tab 0       \tab 1        \tab \code{.} \tab \code{.} \tab \code{.} \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
4       \tab 0       \tab 1        \tab \code{ } \tab \code{} \tab \code{ } \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
5       \tab 0       \tab 1        \tab \code{.} \tab \code{.} \tab \code{.} \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
6       \tab 0       \tab 1        \tab \code{ } \tab \code{} \tab \code{ } \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
7       \tab 0       \tab 1        \tab \code{.} \tab \code{.} \tab \code{.} \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
8       \tab 0       \tab 1        \tab \code{ } \tab \code{} \tab \code{ } \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
9       \tab 0       \tab 1        \tab \code{.} \tab \code{.} \tab \code{.} \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
10      \tab 0       \tab 1        \tab \code{ } \tab \code{} \tab \code{ } \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
\code{.}\tab \code{.}\tab \code{.} \tab \code{.} \tab \code{} \tab \code{ } \tab \code{.}  \tab \code{.}\tab \code{.} \tab \code{.} \tab \code{.}\cr
\code{.}\tab \code{.}\tab \code{.} \tab \code{ } \tab \code{.}\tab \code{ } \tab \code{.}  \tab \code{.}\tab \code{.} \tab \code{.} \tab \code{.}\cr
\code{.}\tab \code{.}\tab \code{.} \tab \code{} \tab \code{} \tab \code{.} \tab \code{.}  \tab \code{.}\tab \code{.} \tab \code{.} \tab \code{.}\cr
312     \tab 1       \tab 1        \tab \code{.} \tab \code{.} \tab \code{.} \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
313     \tab 1       \tab 1        \tab \code{} \tab \code{} \tab \code{ } \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
314     \tab 1       \tab 1        \tab \code{.} \tab \code{.} \tab \code{.} \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
315     \tab 1       \tab 1        \tab \code{ } \tab \code{} \tab \code{ } \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
316     \tab 1       \tab 1        \tab \code{.} \tab \code{.}\tab \code{.} \tab 0    \tab 0    \tab 0    \tab 1    \tab 0    \cr
 }
}
}
\keyword{datasets}
