% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDatasets.R
\docType{data}
\name{Respiratory}
\alias{Respiratory}
\title{Patient's respiratory staus data}
\format{
\describe{
   \item{subject}{Number of patients}
   \item{gender}{Patient gender ( 0 : Female, 1 : Male )}
   \item{age}{Patient age ( 0 : Under 30, 1 : Over 30 )}
   \item{month}{Measurement time ( 0 : before, 1 : 1 month, 2 : 2months )}
   \item{status}{Measurement status after taking placebo, response variable ( 0 : poor, 1 : good )}
}
}
\source{
Davis. (1977).
}
\usage{
data(Respiratory)
}
\description{
This is part of the data on the patient's respiratory status.
}
\details{
{
57 patients were measured for good and bad by taking 3 measurements before, 1 and 2 months after taking placebo
A data frame with 171 rows and 5 variables \cr
\tabular{rccccc}{
\strong{ } \tab  \strong{subject}  \tab \strong{gender}  \tab \strong{age}  \tab \strong{month}  \tab \strong{status}\cr
1        \tab 1        \tab 0    \tab 1        \tab 0    \tab 0    \cr
2        \tab 1        \tab 0    \tab 1        \tab 1    \tab 0    \cr
3        \tab 1        \tab 0    \tab 1        \tab 2    \tab 0    \cr
4        \tab 2        \tab 0    \tab 0        \tab 0    \tab 0    \cr
5        \tab 2        \tab 0    \tab 0        \tab 1    \tab 0    \cr
6        \tab 2        \tab 0    \tab 0        \tab 2    \tab 0    \cr
7        \tab 4        \tab 0    \tab 1        \tab 0    \tab 1    \cr
8        \tab 4        \tab 0    \tab 1        \tab 1    \tab 1    \cr
9        \tab 4        \tab 0    \tab 1        \tab 2    \tab 1    \cr
\code{.} \tab\code{.}  \tab\code{.}\tab \code{.}\tab\code{.}\tab\code{.}\cr
\code{.} \tab\code{.}  \tab\code{.}\tab \code{.}\tab\code{.}\tab\code{.}\cr
\code{.} \tab\code{.}  \tab\code{.}\tab \code{.}\tab\code{.}\tab\code{.}\cr
166      \tab 104      \tab 1    \tab 1        \tab 0    \tab 1    \cr
167      \tab 104      \tab 1    \tab 1        \tab 1    \tab 0    \cr
168      \tab 104      \tab 1    \tab 1        \tab 2    \tab 1    \cr
169      \tab 106      \tab 1    \tab 1        \tab 0    \tab 0    \cr
170      \tab 106      \tab 1    \tab 1        \tab 1    \tab 0    \cr
171      \tab 106      \tab 1    \tab 1        \tab 2    \tab 0    \cr
 }
}
}
\references{
Davis, C. S. (1991). Semi-parametric and non-parametric methods for the analysis of repeated measurements with applications to clinical trials. \emph{Statistics in Medicine}, \bold{10}, 1959—1980.
}
\keyword{datasets}
