% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlda_add.R
\name{vlda_add}
\alias{vlda_add}
\title{Supplementary Objects and Variables}
\usage{
vlda_add(fit, add.col = NULL, add.row = NULL, time.name = NULL)
}
\arguments{
\item{fit}{An object returned by vlda()}

\item{add.col}{A data matirx, The type of indicator matrix. Additional data sets in column format. \eqn{p \ge 2}}

\item{add.row}{A data matirx, The type of indicator matrix. Additional data sets in row format.
Supplemental data should have the same variable name as \code{fit$ind.mat} returned by \code{\link{vlda}},
and if it is not an indicator matrix, you can use it after generate an indicator matrix using \code{\link{indicator}} built into vlda.}

\item{time.name}{If supplemental data to add contains a time variable, it requires argument a character string that specifies the name of the time variable.}
}
\value{
\item{...}{Same as the result of vlda}
\item{sup.coordiante}{A tibble data class, coordinates of new objects created when adding supplementary data in the vlda plot already provided.}
}
\description{
Add objects or variables with new information to the two-dimensional VLDA plot proposed for multidimensional longitudinal data.
}
\details{
The longitudinal data inevitably has the characteristic that supplementary data is added.
Supplement data is outcome variables measured at additional time points, such as  after the last time point ,
or new objects not previously measured, can be added to the row in the long form.
And in addition to the previously given covariates, other variables that characteristics of the object can be added
to the columns. Therefore, find coordinates representing objects and variables that are added in the VLDA plot already
provided, through a method obtain that find coordinates on low-dimensional space for supplementary elements.
}
\examples{
#### Supplementary row and column indicator matrix added ####
### long form ###
data(PTSD)
PTSD <- as.data.frame(PTSD)
PTSD[,2:4] <- apply(PTSD[,2:4], 2, function(x) ifelse(x >= 3, 1, 0))
PTSD[,5] <-  ifelse(PTSD[,5] >= 6 , 1, 0)
PTSD <- data.frame(lapply(PTSD, function(x) as.factor(x)))
fit <- vlda(x = PTSD, object = "subject", time = "time", type = "long")

## Column indicator matrix added ##
# The degree of drinking that may affect PTSD
data(PTSD_column)
PTSD_column <- as.matrix(PTSD_column)
vlda_add(fit, add.col = PTSD_column )


## Row indicator matrix added ##
# Added to the row, and is intended for 316 patients after 18 months.
# (The value of the time variables is zero vector)
data(PTSD_row)
PTSD_row <- as.matrix(PTSD_row)
vlda_add(fit, add.row = PTSD_row)


## Row and Column added ##
vlda_add(
fit,
add.row = PTSD_row,
add.col = PTSD_column
)


### Wide form ###
data(Depression)
fit2 <- vlda(x = Depression, object = "Case", time = c("1week", "2weeks", "4weeks"), type = "wide")

## Column indicator matrix added ##
# Response after 6 weeks and gender were added the columns for 800 existing patients.
data(Depression_column)
Depression_column <- as.matrix(Depression_column)

## Row indicator matrix added ##
# 100 patients who took placebo in each group of mild and severe were added to the rows.
# (The value of the drug dummy variable is zero vector)
data(Depression_row)
Depression_row <- as.matrix(Depression_row)


## Row and Column added ##
vlda_add(
fit2,
time.name = "6weeks",
add.row = Depression_row,
add.col = Depression_column
)


}
\seealso{
\code{vlda}
}
\keyword{Supplement}
